requirements.txt
torch
torchvision
fastapi
uvicorn
python-multipart
pillow
scikit-learn
matplotlib
seaborn
wandb
numpy